numeric: T4467: add support for relative values: +/-N
Add option '--relative' to check that value is of the form '+/-N' for
some numeric N (e.g. +10' or '-30'), interpreted as a relative increment
or decrement, rather than a signed integer/float.  This may be combined
with any other option, which will be applied to the value stripped of
the leading modifier.

Testing this extension revealed that negative values are being
interpreted as options, whether quoted or not; to address that, an
Arg.Rest keyword '--' is introduced, which will need to be passed along
with the option in interface definitions using the above. For example:
<validator name="numeric" argument="--relative --"/>